requirements.txt
numpy==2.0.0
opencv-python==4.10.0.84
pillow==10.3.0
